Removed +server.ts functionality
